MASSIVE REFACTOR
https://youtu.be/tlCFQ5GTe3w
Rewrote much of the game's code to properly use inheritance, interfaces, delegates/callbacks, singletons, Unity scriptable objects, and to generally move things towards and MVC pattern.
There are now five resources: Food, Wood, Stone, Gold, Mana. Inspired, huh? It'll do for now.
Structures can have multiple abilities. This was one of the main motivations behind the refactor. For now, ability means attack, although that wont always be the case!
Structures have descriptions, displayed when they are selected after being built.
In defend mode, each of a struture's abilities, along with the cooldown, is displayed in the bottom left when the structure is selected.