Units and Build Mode Abilites
Added units. Each tile can contain one unit and one structure. Units cannot be built but start in the level.
Added build mode abilities. These are abilities such as unit movement that can be executed in build mode for a resource cost.
Added a Wizard unit that can move and attack with the Alchemist Bomb AOE attack, and a Peasant unit that can move and do absolutely nothing else!
Created a Level scriptableobject and moved a lot of level-specific data within it, such as starting units, structures and resources, and waves, making it much easier to create and tweak levels.
Removed debug spawning (creeps can only be spawned in normal waves)