Basic xp/fatigue tracking and new tiles
Added tracking of how much time a unit spends with at least one defend mode ability on cooldown, which gives the experience and fatigue for that defend mode round. Neither experience or fatigue are used yet, just calculated and stored
Created a HexType ScriptableObject to hold information about each type of hex separately to the information about individual hexes within HexData, making it easier to create and use different types of hex
Based on this, there are now five hex types: grassland, forest, mountain, desert and water
What was the HexType enum is now BaseHexTypeEnum - it's there so that hex types are drop-down selectable within the unity editor. It's a bit ineligant and may not be needed but it's there for now
Added a new, more interesting level using the above changes