Unit Levelling and Talents, Editor UI Toggler
Units can now level up based on passing experience thresholds.
Fatigue now affects experience. There is maths involved in this.
Added a system of Unit Attributes, Attribute Modifiers and Unit Talents. The way I've coded this is definitely in the wait-and-see-if-its-actually-good-or-not category.
Unit Talents contain Attribute Modifiers, which will be added to the unit when a Talent is levelled up
Added UI to view Unit Talents and upgrade them if the unit has a level up point available.
Added a fairly bad editor script that allows Build mode and Defend mode ui to be switched out in the editor. It's shite but the whole of the UI will be redone at some point anyway.