Initial version of Auras
Units and Structures can now have auras, which can apply improvements (basically any kind of positive effect, but normally attribute bonuses like increased range) to other units and/or structures within a defined range.
Moved improvements to an interface, which the relevant ScriptableObject implements. Will probably roll out this patters accross all SOs because it means that a non SO version can be generated in code in the standard way, which amongst other things means they can be unit tested, and also I'm not tied to SOs if I decide there's a better implementation
Added CallbackList class, which is a list that has a callback when you add and remove stuff :)