Economy Refactor
Refactored Resources to be So/Is, so that the model code is agnostic to the number and names of resources in the game, and so that the economy uses the same basic pattern as hex types etc now use.
Refactored classes like EconomyTransaction and the EconomyManager for this change.
Still bits to do around the edge of this but the core is now ok.