Creep Attributes and Generic Temporary Effects
Creeps have attributes rather than hardcoded fields like speed. Left hitpoints as a field for now as it's probably enough of a special case?
PersistentEffects renamed to TemporaryEffects, made more robust and moved out of Creep for cleanness and future composition.
More (yes more) AttackEffectType refactors related to the above - non-damage types are now attribute modifiers and resistance modifiers (there may be others, if more special cases like resistances arise). These can both be permanent or temporary. When a creep receives them an attribute modifier is created to affect the change.
There must be bugs as of this commit but the game compiles and runs, and slows still work, so it's pretty much there.