C# events and auto implemented properties
Converted all callback Actions to use C# events. Did this when I tried to implement what was basically the same thing and realised the language already had a mechanism for it. There are reasons I'm not thrilled by events, EventHandlers and EventArgs, but it's probably better than the pattern I had before, albeit probably not better enough that it would have been worth converting if I were working on this as something to ship quickly rather than something to play with.
Converted all explicit get-only public properties with private backing members in concrete (i.e. non-serializable) classes to auto implemented properties. I think the only impact is (significantly) shorted and clearer code. I believe I can't do this on serialized classed (or at least the SerializedFields in them) because Unity needs the explict member to serialize.
Fixed not incrementing the wave number (and so not spawning any waves after the first)