First unit tests, build mode ability refactor
Added the first unit tests, and in doing so refactored build mode abilities somewhat. The new pattern is more complex, possibly overly so, and I'm worried that I'm generalising and abstracting before I have use cases that warrant it. However the new pattern is unit testable, which the other one wasn't so much.