Refactor DefendModeAbilityController to model object
Finally killed DefendModeAbilityController as a MonoBehaviour and moved it to a ModelObjectFrameUpdatee manager class attached to CAbilities.
Refactored this manager code to be more event driven, including changing DefendModeAbilities themselves to emit events.
Moved unit time tracking inside CUnit itself as event responses and ModelObjectFrameUpdates instead of the old hacky crap in the manager/controller
Sketched out unit tests for this stuff